+++ site/skater_backend/src/main/java/com/ecommerce/skater/repository/PaymentMethodRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PaymentMethodRepo extends JpaRepository<PaymentMethod, Long> {
+public interface PaymentMethodRepo extends JpaRepository<PaymentMethod, Integer> {
 }
