+++ site/skater_backend/src/main/java/com/ecommerce/skater/repository/ProductImageRepo.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ProductImageRepo extends JpaRepository<ProductImage, Long> {
+public interface ProductImageRepo extends JpaRepository<ProductImage, Integer> {
 
-    List<ProductImage> findAllByProductId(Long id);
+    List<ProductImage> findAllByProductId(int id);
 }
