+++ site/skater_backend/src/main/java/com/ecommerce/skater/repository/ShipmentRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ShipmentRepo extends JpaRepository<Shipment, Long> {
+public interface ShipmentRepo extends JpaRepository<Shipment, Integer> {
 }
